plumbing to make timestampcache live re-size-able
