+++ atlasdb-client/src/main/java/com/palantir/atlasdb/transaction/impl/AbstractTransactionManager.java
@@ -29,7 +29,6 @@ import org.slf4j.LoggerFactory;
 import com.google.common.base.Preconditions;
 import com.google.common.util.concurrent.RateLimiter;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import com.palantir.atlasdb.cache.TimestampCache;
 import com.palantir.atlasdb.transaction.api.Transaction;
 import com.palantir.atlasdb.transaction.api.TransactionFailedException;
 import com.palantir.atlasdb.transaction.api.TransactionManager;
@@ -42,7 +41,6 @@ public abstract class AbstractTransactionManager implements TransactionManager {
     private static final int GET_RANGES_QUEUE_SIZE_WARNING_THRESHOLD = 1000;
 
     public static final Logger log = LoggerFactory.getLogger(AbstractTransactionManager.class);
-    protected final TimestampCache timestampValidationReadCache = new TimestampCache();
     private volatile boolean closed = false;
 
     @Override
@@ -125,11 +123,6 @@ public abstract class AbstractTransactionManager implements TransactionManager {
      */
     protected void checkOpen() {
         Preconditions.checkState(!this.closed, "Operations cannot be performed on closed TransactionManager.");
-    }
-
-    @Override
-    public void clearTimestampCache() {
-        timestampValidationReadCache.clear();
     }
 
     ExecutorService createGetRangesExecutor(int numThreads) {
