+++ atlasdb-config/src/main/java/com/palantir/atlasdb/config/AtlasDbRuntimeConfig.java
@@ -20,6 +20,7 @@ import org.immutables.value.Value;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.palantir.atlasdb.AtlasDbConstants;
 
 @JsonDeserialize(as = ImmutableAtlasDbRuntimeConfig.class)
 @JsonSerialize(as = ImmutableAtlasDbRuntimeConfig.class)
@@ -42,6 +43,19 @@ public abstract class AtlasDbRuntimeConfig {
     @Value.Default
     public TransactionConfig transaction() {
         return ImmutableTransactionConfig.builder().build();
+    }
+
+    /**
+     * The number of timestamps to cache that we have seen in previous reads.
+     * This will use somewhere around 90MB of heap memory per million timestamps because of various overheads
+     * from Java Objects and the cache's LRU tracking.
+     *
+     * Probably the only reason to configure away from the default would be a service that can afford the heap usage,
+     * and has read patterns that deal with a very large working set of existing transactions.
+     */
+    @Value.Default
+    public long getTimestampCacheSize() {
+        return AtlasDbConstants.DEFAULT_TIMESTAMP_CACHE_SIZE;
     }
 
     public static ImmutableAtlasDbRuntimeConfig defaultRuntimeConfig() {
